Move codeblock and fileblock to UI/Display
