+++ src/components/GettingStarted/Steps/SimpleExample.js
@@ -8,9 +8,8 @@ import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import { OrganizationsRoutes } from 'shared/constants/routes';
 import Button from 'UI/Controls/Button';
+import { CodeBlock, Output, Prompt } from 'UI/Display/CodeBlock/CodeBlock';
 import Aside from 'UI/Layout/Aside';
-
-import { CodeBlock, Output, Prompt } from '../CodeBlock';
 
 class SimpleExample extends React.Component {
   state = {
